feature: replace `TxStatus` with `SubmittableResult` (#250)
instead of TxStatus, which communicated transaction status with its `status` property, we are now using the class that the substrate sdk is returning on transaction submissions, which is a `SubmittableResult`. This has the two properties `isFinalized` and `isError` that communicate the status much better, and has more such properties on the `ExtrinsicStatus` Enum it stores on its `status` property.