explicit transactions & awaiting custom tx status changes (#299)
Functions that require blockchain writes will no longer implicitly submit the transaction. A signed SubmittableExtrinsic is returned instead. The user is required to explicitly trigger transactions by calling its `send()` method or by passing it to the `submitSignedTx()` helper introduced with this change. The latter can be configured to await different states along the transactions lifetime and takes care of listening to error events.